[IMP] base, *: allow html fields not to sanitize forms
* = website_slides

The option to sanitize or not the forms was not available, this will
allow better flexibility on whether forms should be sanitized or not on
an HTML field.
Also we use this new param to allow forms to be added on an already
existing html field where the form was sanitized out.

This is a backport of 388c222c6c4bb7e2fe3e67009b248359ae0fd3db
adapted to the needs in 13.0

closes odoo/odoo#77266

Signed-off-by: Xavier Morel (xmo) <xmo@odoo.com>