[FIX] purchase_stock: set propagate_date according to buy rule
Before this commit, when we manually create a PO line, the
propagate_date will always be false. So the change on scheduled date
will never pass to corresponding move.

After this commit, when create a PO line, if no propagate_date for it,
we find if the product has a route with a buy rule. If yes, use the
setting on its first buy rule, if no, set the default to be true.

Task 2298950
PR #55155

closes odoo/odoo#94272

Signed-off-by: William Henrotin (whe) <whe@odoo.com>