[FIX] stock_picking_batch: remove empty pickings
When validating a batch picking where some pickings have no quantities
done + some do, then avoid immediate transfers wizard and remove
pickings with no quantities done from the batch picking. This will make
it so users can validate a partially done batch picking in this case.

This is an adaptation of [1] (which is already deployed in 15.0)

[1] eb0ebc077afed24e8f5c2212f2a216bb1b8e6936

OPW-2711585

closes odoo/odoo#81844

Signed-off-by: Tiffany Chang <tic@odoo.com>
Co-authored-by: William Henrotin (whe) <whe@odoo.com>