[FIX] web_editor, website: fix removal of last item inside active slides
A bug currently makes the carousel snippet collapse/disappear when its
active contents are removed. This happens because after the removal,
the carousel no longer has an active item.

The solution that was chosen was to not allow the removal of slides
unless using the dedicated option which is there for that. This solution
has to advantage to also allow users to create empty slides (using a
background image for example) by simply removing the columns inside.

task-2506165

closes odoo/odoo#77033

Signed-off-by: Quentin Smetz (qsm) <qsm@odoo.com>