[FIX] website: correctly display overlay on a new slide
Before this commit, selecting elements on a newly added slide of the
carousel was not displaying the editor overlay.
It was due to the fact that adding a new slide to the carousel was done
making a deep clone of the slide, and therefore of its editor.
Now, the slide is cloned without its editor.

task-2657532

closes odoo/odoo#77335

Signed-off-by: Quentin Smetz (qsm) <qsm@odoo.com>