[FIX] event_sale: create tickets with default product type "service"
In some cases when creating a new ticket, the default product
type shown is a "Consumable". It should rather be handled as a
"Service". In later versions, a detailed type is created, but the
base product type should be set to service too.

This commit uses the `product_id` `context` to set the default
type to "service".

Task-2655226

closes odoo/odoo#78046

Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>