[FIX] account: remove readonly on auto_post
The aim of this commit is to allow user to post a reversed entry even
when the date for the reversed has been set in the future.

before this commit:
if the entry move is generated using the reverse entry button and a
date in the future, the reverse entry is created with auto_post True
and is readonly in the view resulting in the user being unable to post
the move himself.

after this commit:
auto_post can be manually set to false and the user can post the move
himself.

closes odoo/odoo#81253

Task: #2522640
Signed-off-by: Olivier Colson <oco@odoo.com>