[FIX] pos_cash_rounding: Fix value under rounding
It should be allowed to pay amount that are not rounded if the value is less than the rounding value.
Ex:
Rounding value: 0.05
Amount to pay: 0.04
The amount to pay should remain 0.04 and not be rounded to 0.05.

This fix allows the user to pay for values under the rounding value without applying the rounding.

closes odoo/odoo#81302

Signed-off-by: Masereel Pierre <pim@odoo.com>