[FIX] l10n_lu: fix tax report formula for 012 and 454
- Configure LU localization
- Go to Accounting > Accounting > Journal Entries
- Create a new entry with the following items:

          Account        | Taxes  | Debit | Credit | Tax Grids
 --------------------------------------------------------------
 746000 Benefits in kind | 17-ATN |   0 € |  100 € | (456)(701)
 461411 VAT received     |        |   0 € |   17 € | (702)
 621150 Benefits in kind |        | 117 € |    0 € |

- Post it
- Go to Accounting > Reporting > Tax Report
Lines "I.A. Overall turnover (012)" and "I.A.2. Total Sales / Receipts (454)"
are incorrectly computed and have an exceeding amount of 100€.
Their child line "I.A.2.b) Other sales / receipts (472)" has the following
formula: LUTAX_021 + LUTAX_037 - LUTAX_456
But the substraction part of this formula (- LUTAX_456) is not present in
their own formula.

LUTAX_456 should be substracted from the formula of lines (012) and (454).

opw-2699920

closes odoo/odoo#82162

Signed-off-by: William André (wan) <wan@odoo.com>