[FIX] website_sale_comparison: Ensure product_data available
Check whether the product ID of the first item exists inside product_data
before trying to access it. Although it means the first item still does not
appear, both products will be present in the user's cookies.
When the user refreshes the page or navigates away, both products will be
visible in the comparison panel.

closes odoo/odoo#64259

Signed-off-by: William Braeckman (wbr) <wbr@odoo.com>