[IMP] account: search on stored parent_state instead of related
The parent_state is a stored field on account.move.line while the move_id relates to another model.
By using the parent_state (stored) instead of the relation we save SQL joins.
Besides of this we benefit from the index on company_id and parent_state to optimize some queries.

closes odoo/odoo#82986

Signed-off-by: Quentin De Paoli <qdp@odoo.com>