[FIX] l10n_fr_fec: test shouldn't be dependent on the date
The test was failing because of the hardcoded date in the expected test
values, whereas the real values gotten were using today's date.

The test should also be running when installing only `l10n_fr_fec`, so
we should be installing a chart template from `l10n_fr`, not the generic
one.

The field `test_file` didn't exist in this version and was probably
added in a later version and wrongly backported.

closes odoo/odoo#83660

Signed-off-by: Florian Gilbert <flg@odoo.com>