[REV] account: commit d22cd5473326b3b5af4ae44b0d121349a1a90850
d22cd5473326b3b5af4ae44b0d121349a1a90850 introduced the fact that when doing a full refund of an invoice, the same account was used as on the original invoice, instead of the one on the tax refund repartition lines, though the tags were still the refund ones. This is wrong.

Also, it restored an inconsistent data setup in a test, while the current version was actually correct.

Part-of: odoo/odoo#83837