[FIX] sale_coupon: discount applied on the price exc. taxes
If applied, this commit will fix the following bug by using the correct
total based on whether the price is tax inc. or exc.

Steps to reproduce:
- Install sale
- Create a tax with "included in price" set on it
- Create a promotion program with any value higher than
  [100/(1+VAT/100)]% discount on total order
- Create a SO and add product with the created tax
- Add promotion program

Result:
- The discount applied is capped at [100/(1+VAT/100)]%
  Eg, for a 35% included tax, the max discount will be 74.07%,
  regardless if the discount % is higher like 90%.
Expected result:
- The correct discount should be applied

Bug:
The ```price_subtotal``` is used to calculate the order total amount
whether the price is tax inc. or tax exc.

Fix:
Use ```price_total``` when the price is tax inc.

opw-2765883

closes odoo/odoo#85955

Signed-off-by: Romain Derie (rde) <rde@odoo.com>
Co-authored-by: Andrea Grazioso (agr-odoo) <agr@odoo.com>