[FIX] account_payment : limit token acquirers to invoice's partner
Steps:
Log in with admin.
Configure Stripe for direct payment (from Odoo).
Create an invoice for portal.
Click on Preview, then Pay Now.
Select Stripe, enter your credentials and pay.

Issue:
Error message: Invalid token found!

Cause:
Paying this way creates a token for admin.
Yet, later, we check that invoice's partner == token's.
Using acquirers with a token-based payment method
should not be allowed when the current user's partner
is not the invoice's customer.

Fix:
Clarify the error message to indicate how to resolve it :
log in as the good partner.

opw-2730967

closes odoo/odoo#88253

Signed-off-by: Antoine Vandevenne (anv) <anv@odoo.com>