[FIX] website_hr_recruitment: allow portal users to filter by department
Currently, only a public user or an internal user can access HR
department information. This is causing an access rights error when
trying to filter job offers on the website if connected with a portal
user, while there is no error if not connected (public user). We should
have the same behavior for the two cases.

closes odoo/odoo#91733

Signed-off-by: Romain Derie (rde) <rde@odoo.com>