[FIX] mrp: correct logging of quantity exception note on pickings
It could happen that there are other objects than stock.move linked to
the manufacture order. This is currently creating an error when
concatenating the objects. As we are only keeping stock.move objects to
log the exception on those, we should filter them before the concatenate
operation to avoid the error.

closes odoo/odoo#94830

Signed-off-by: Arnold Moyaux (arm) <arm@odoo.com>
Co-authored-by: vancleeff <mvw@odoo.com>