[FIX] hr_holidays: Properly make 'My Time Off' action buttons invisible
Since the action buttons are not field nodes, the 'invisible' attribute
doesn't do anything. The only way to make them invisible is to have an
'invisible' key in attrs/modifiers.

The 'states' attribute was removed, since ir_ui_view overwrites the
'invisible' item in the modifiers dict if 'states' exists in the node.

closes odoo/odoo#147180

X-original-commit: d0b19749dcd88e982db2a26edcfd2efd27505ac4
Signed-off-by: Sofie Gvaladze (sgv) <sgv@odoo.com>