[FIX] l10n_de: remove tag B from ir_model_data
In odoo/odoo/pull/126249 the german balance sheet report was updated and tag tag_de_liabilities_bs_B was removed from account_account_tags_data.xml. This deletes it from ir_model_data so that the ORM doesn't delete this record. This is done so that it doesn't break if another table was already using this tag.

closes odoo/odoo#131530

Signed-off-by: Laurent Smet (las) <las@odoo.com>