[REV] l10n_de: remove l10n_de_document_title field from template"
This reverts commit 95468505f0bf1e71bcfb6ab73d893f0a1b9e4dc5.

The commits break the preview, as docs is not defined in preview (while done in real printing).
Not sure what we will do, with a check on the presence of the field or not, as it was working fine before, only a wrong modification of a view made it visible.
Linked to runbot error 24276

closes odoo/odoo#132639

Signed-off-by: Florian Gilbert (flg) <flg@odoo.com>