[FIX] web_editor: avoid adding a font tag after an empty p within a td
When applying font color to a completely selected `td` containing  a empty `p`
tag, it results in a creation of a `font` tag after the `p` tag, as it checked
whether the `node.parentNode` is `EmptyBlock`.  Alongside with this, it should
also check if the node is 'BR'.

Part-of: odoo/odoo#153217