[FIX] web: clipped dropdown on firefox in list views
On firefox and safari applying a position-sticky on an element inside
an other element with an overflow is causing a rendering issue.

This was occuring with the dropdown `.o_optional_columns_dropdown`
making it unusable as soon as the `overflow-x` triggered the horizontal
scroll in the table.

This commit moves the `o_optional_columns_dropdown` out of the natural 
DOM flow with a position-absolute, removing the position-relative on 
each parents until the `.o_field_x2many.o_field_x2many_list` which is 
out of the horizontal scroll applied by `table-responsive`. This ensures
the dropdown-toggle is always displayed on the top right of the table.

task-3696473

closes odoo/odoo#151570

Signed-off-by: Pierre Paridans (app) <app@odoo.com>