[FIX] html_editor: fix link plugin popup in firefox
**Steps to reproduce:**
(Firefox only)
- Go to any chatter
- Open the full composer
- Try to add a link
(using `/link` or by selecting existing text and using the toolbar)
- Change a value in the popup
- On each input the cursor gets back to the editor
and you need to manually move it back to type again

**Issue:**
When inside a modal (Full Composer), the autofocus
hook in the `LinkPopover` setup does not behave as
expected.

The active element is assigned to the modal container
instead of the input element inside the popover, which
does not receive focus.

**Fix:**
Backporting autofocus fix from [1].

[1] https://github.com/odoo/odoo/commit/927fa045e1565a39ef6cb2c9b241686c62c234fe

opw-6344131

closes odoo/odoo#276468

Signed-off-by: David Monjoie (dmo) <dmo@odoo.com>