[FIX] stock_account: rounding error on revaluation
Before this commit, the remaining_value_unit_cost was
rounded before any computation.
In the case where the numer of layers with remaining value
and remaining quantity increase, the rounding error introduced
by that rounding quickly explodes, leading to a
negative remaining_value during revaluation computation.

After this commit, the remaining value is rounded at the end, after
the computations and the checks.
This ensures that the rounding error remains constant and does not
accumulate over the execution of the method.

opw-4901966

closes odoo/odoo#228767

X-original-commit: fc06e7c249bb8df8d8f28f26d3c8d2dfc197eab2
Signed-off-by: Quentin Wolfs (quwo) <quwo@odoo.com>
Signed-off-by: Victorien Sonneville (viso) <viso@odoo.com>