[FIX] pos_loyalty: keep coupon custom codes
**Steps to reproduce:**

- Go to PoS
- Buy a gift card, click on the sell physical gift card
- Set a custom code and price
- Reload the page
- Go through with the payment and check the downloaded pdf
- The coupon code is an automatically generated one

**Problem:**
When reloading the coupons after reloading the order, the
gift cards lose their custom codes if it was set before.
This means that the coupon will have an automatically
generated name in the backend as well as in the pdf that is
downloaded once the payment is done.

**Why the fix:**
This happens because we are trying to access the
order.get_selected_orderline()?.gift_code to get the
coupon code.
https://github.com/odoo/odoo/blob/f13ee41a4a62f7fedbddc4245e52ee8e9dce9e0e/addons/pos_loyalty/static/src/overrides/models/pos_store.js#L230
But when we leave the order to go back in afterwards,
there is no selected line anymore, so the gift_code is
undefined. We then fallback to the default code, which
is to generate one for the user.

The gift_code is still on the orderline, so all we need
to do is to link the newly created coupon with the
gift_code in the orderline.

Another problem that arises her is that when purchasing
a custom gift card AND a default gift card in the same order,
then reloading, the gift codes will be mixed up in the pdf and
in the backend as well.
This happens because the custom code's name in the backend is 'code'
and we use gift_code in the frontend, so we assign the gift_code to
the code in the backend.

opw-4964509

closes odoo/odoo#223442

Signed-off-by: Stéphane Vanmeerhaeghe (stva) <stva@odoo.com>