[FIX] marketing_card: raise on failed image render
If the wkhtmltoimage fails for some reason we currently keep going as if
an image was actually rendered. Instead if the result of the image
render is `None`, raise a generic error.

This avoids issues with cards being marked "synced" even though they are
actually empty.

Also update the "testing" check to be simpler, and remove the useless
"force rendering" context keys. Rendering results don't need unit tests.

task-5048534

Part-of: odoo/odoo#224629
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>