[FIX] website: avoid page refresh after deleting search
Before this commit, clicking on the "x" button on the searchbar would
refresh the page and remove the result immediately. The user might want
to keep the result if he does not do another search.

This commit avoids refreshing the page when removing the input content.

Steps to reproduce the bug:
- Click on the search button in the navbar
- Search "tree"
- Press enter
(You are redirected to /website/search)
- Click on the "x" button to delete the input content
(The page refreshes)

task-3570756

closes odoo/odoo#231322

X-original-commit: 164877d0ce138da6aaf788a5dedb9635f5fc8e2e
Signed-off-by: Serge Bayet (seba) <seba@odoo.com>
Signed-off-by: Romaric Moyeuvre (romo) <romo@odoo.com>