[FIX] hr_timesheet: show allocated hours on tasks even if equals to 0
Before this commit, when the user would like to edit the allocated hours
field of multiple tasks inside the list view, he cannot when the
allocated hours is equal to 0. The reason is because the cell is hidden
once the value is equal to 0.

This commit makes that field is always editable in the list view of
tasks even if the value is 0.

task-5093288

closes odoo/odoo#228862

X-original-commit: 089aa4e0839d9ce76c6664e3498b6ce9de1aaeb6
Signed-off-by: Vincent Larcin (vila) <vila@odoo.com>
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>