[FIX] base: ir_qweb_field: Handle One2Many html converter
For now, one2many fields are not displayed correctly in reports.
This is because for other types of fields (ManyToMany, ManyToOne, etc.),
there are Converter classes (ManyToManyConverter) “ir.qweb.field.many2many”
that implement `value_to_html`.
However, for OneToMany fields, no implementation is present.

This commit adds a `OneToManyConverter`, which is essentially the same
as the ManyToMany one.

opw-5098466

closes odoo/odoo#230539

X-original-commit: 575d3625a2fc7dc84af0403388304aba4a296727
Signed-off-by: Achraf Ben Azzouz (abz) <abz@odoo.com>
Signed-off-by: Christophe Simonis (chs) <chs@odoo.com>