[IMP] point_of_sale: disable autofocus on pos config kanban for mobiles
Task: [#5016943](https://www.odoo.com/odoo/project/1737/tasks/5016943)

---

On tablets and phones, the search bar was autofocus when opening the POS config kanban view.
This was causing the keyboard to open automatically, which was not a good user experience.

Now the search bar is not autofocus on touch devices for the POS config kanban view ('view_pos_config_kanban').

closes odoo/odoo#231045

Related: odoo/enterprise#96873
Signed-off-by: David Monnom (moda) <moda@odoo.com>