[FIX] website_blog: add missing demo data for blog_tags_with_date tour
With PR [1], the blog_tags_with_date tour was introduced, but the
required demo data were missing, causing the tour to fail when executed
in an environment without demo data.

This commit adds the necessary demo records to ensure the tour runs
successfully even when no demo data are present.

[1]: https://github.com/odoo/odoo/pull/225845

closes odoo/odoo#231202

X-original-commit: 747f775759b293fbc5cd7652739d94c6876d57e9
Signed-off-by: Serge Bayet (seba) <seba@odoo.com>
Signed-off-by: Divyesh Vyas (divy) <divy@odoo.com>