[FIX] purchase: prevent deletion of UoM when it is in use
When a user deletes the UoM used in a purchase order line and then tries to
confirm the purchase order.

Steps to reproduce:
- Install 'purchase_stock' module(without demo)
- Create a New PO > Add a product in Line (with UoM=Units)
- Remove UoM in order line and select 'Dozen' in it > Save
- Settings > Units of Measure Categories > Open 'Units' > Remove 'Dozen'
- Orders  > Requests for Quotation > Open PO > 'Confirm Order'

Traceback:
ValueError: Expected singleton: uom.uom()
AssertionError: precision_rounding must be positive, got 0.0

This error occurs because, after the UoM is deleted, the 'product_uom' field
becomes empty, which leads to an error.

Solution:
This commit resolves the error by restricting the deletion of a UoM when it is
still in use.

sentry-6746792383

closes odoo/odoo#231478

Signed-off-by: Arnold Moyaux (arm) <arm@odoo.com>