[FIX] sale: don't skip super on validation post-process
Versions
--------
- 18.0+

Issue
-----
`super` isn't called on `_post_process` for confirmed validation
transactions.

Cause
-----
A fix that prevented the sending of mail for validation transactions
added a `continue` in a loop, which skips over the `super` call in the
loop.

Solution
--------
Instead of `continue`, add an `if` check to see if we want to send mail.

closes odoo/odoo#231972

Signed-off-by: Victor Feyens (vfe) <vfe@odoo.com>