[FIX] mass_mailing: Limit name field to it's container
Issue:
When you create a long name for your maling list, the name overflows over the other fields in the view.

Repro:
Go to marketing -> mailing lists -> create a fairly long name -> notice name overflows in the view.

Cause:
Small bootstrap issue, name field not limited to it's container.

Fix:
Added a simple w-100 to limit the field to 100% of it's container.

opw-5153270

closes odoo/odoo#232312

Signed-off-by: Damien Abeloos (abd) <abd@odoo.com>