[FIX] website_payment: ensure currency is loaded before donation click
This fixes a very rare race condition in the `test_01_donation` tour...
but it is actually surprising it is not more frequent than it is.

Steps to reproduce:
- Simulate a slow network (e.g. using Chrome DevTools)
- Load a page with a donation snippet
- Quickly fill in the donation form and click the submit button
=> You'll sometimes get a crash.

Or a more precise one:
- Add a 1 minute delay to the `/website/get_current_currency` route
- Load a page with a donation snippet
- Fill in the donation form and click the submit button once the page is
  fully loaded.
=> You'll definitely get a crash.

This commit makes that button's handler async-protected, meaning it will
now properly wait for what is needed (the currency) before proceeding,
but it will also display a loading effect for the duration and prevent
double clicking. Once the form is submitted, the loading effect is
removed for stability safety. In master it can probably be improved.

runbot-220885

closes odoo/odoo#232408

Signed-off-by: Quentin Smetz (qsm) <qsm@odoo.com>